+++ client/src/components/App.tsx
@@ -1,6 +1,7 @@
 import { ApolloProvider } from '@apollo/react-hooks';
 import ApolloClient from 'apollo-boost';
 import axios from 'axios';
+import debounce from 'lodash/debounce';
 import 'normalize.css';
 import React, { useEffect, useState } from 'react';
 import {
@@ -21,18 +22,25 @@ import AdminStates from './adminPanel/AdminStates';
 import AdminUsers from './adminPanel/AdminUsers';
 import MountainDetailPage from './mountains/detail';
 import ComparePeakListPage from './peakLists/compare';
-import CompareAllPeaksPage from './peakLists/compare/CompareAllMountains';
 import PeakListDetailPage from './peakLists/detail';
 import PeakListPage from './peakLists/list';
 import Header from './sharedComponents/Header';
 import UserProfile from './users/detail';
 import ListUsersPage from './users/list';
 
+export interface IAppContext {
+  windowWidth: number;
+}
+
 const client = new ApolloClient();
 export const UserContext = React.createContext<User | null>(null);
+export const AppContext = React.createContext<IAppContext>({windowWidth: window.innerWidth});
 
 const App: React.FC = () => {
   const [user, setUser] = useState<User| null>(null);
+  const [windowWidth, setWindowWidth] = useState<number>(window.innerWidth);
+
+  const appContext = {windowWidth};
 
   useEffect(() => {
     const fetchUser = async () => {
@@ -46,12 +54,28 @@ const App: React.FC = () => {
     fetchUser();
   }, []);
 
+  useEffect(() => {
+    const updateWindowWidth = debounce(() => {
+      setWindowWidth(window.innerWidth);
+    }, 500);
+    window.addEventListener('resize', updateWindowWidth);
+    return () => {
+      window.removeEventListener('resize', updateWindowWidth);
+    };
+  }, []);
+
   const userRoutes = (user) ? (
     <>
       <Route exact path={Routes.Lists}
         render={(props) => <PeakListPage {...props} userId={user._id} />}
       />
+      <Route exact path={Routes.ListsWithDetail}
+        render={(props) => <PeakListPage {...props} userId={user._id} />}
+      />
       <Route exact path={Routes.ListDetail}
+        render={(props) => <PeakListDetailPage {...props} userId={user._id} />}
+      />
+      <Route exact path={Routes.ListDetailWithMountainDetail}
         render={(props) => <PeakListDetailPage {...props} userId={user._id} />}
       />
       <Route exact path={Routes.MountainDetail}
@@ -60,14 +84,20 @@ const App: React.FC = () => {
       <Route exact path={Routes.Friends}
         render={(props) => <ListUsersPage {...props} userId={user._id} />}
       />
+      <Route exact path={Routes.FriendsWithProfile}
+        render={(props) => <ListUsersPage {...props} userId={user._id} />}
+      />
       <Route exact path={Routes.UserProfile}
         render={(props) => <UserProfile {...props} userId={user._id} />}
       />
       <Route exact path={Routes.ComparePeakList}
+        render={(props) => <UserProfile {...props} userId={user._id} />}
+      />
+      <Route exact path={Routes.ComparePeakListIsolated}
         render={(props) => <ComparePeakListPage {...props} userId={user._id} />}
       />
       <Route exact path={Routes.CompareAllPeaks}
-        render={(props) => <CompareAllPeaksPage {...props} userId={user._id} />}
+        render={(props) => <ComparePeakListPage {...props} userId={user._id} />}
       />
     </>
   ) : null;
@@ -85,17 +115,19 @@ const App: React.FC = () => {
 
   return (
     <UserContext.Provider value={user}>
-      <ApolloProvider client={client}>
-        <GlobalStyles />
-        <Router>
-          <Root>
-            <Header />
-            {adminRoutes}
-            {userRoutes}
-            <div id={overlayPortalContainerId} />
-          </Root>
-        </Router>
-      </ApolloProvider>
+      <AppContext.Provider value={appContext}>
+        <ApolloProvider client={client}>
+          <GlobalStyles />
+          <Router>
+            <Root>
+              <Header />
+              {adminRoutes}
+              {userRoutes}
+              <div id={overlayPortalContainerId} />
+            </Root>
+          </Router>
+        </ApolloProvider>
+      </AppContext.Provider>
     </UserContext.Provider>
   );
 };
