+++ client/src/components/peakLists/detail/index.tsx
@@ -1,5 +1,3 @@
-import { useQuery } from '@apollo/react-hooks';
-import gql from 'graphql-tag';
 import React from 'react';
 import { RouteComponentProps, withRouter } from 'react-router';
 import {
@@ -7,125 +5,8 @@ import {
   ContentLeftLarge,
   ContentRightSmall,
 } from '../../../styling/Grid';
-import { Mountain, PeakList, Region, State, User } from '../../../types/graphQLTypes';
-import Header from './Header';
-import MountainTable from './MountainTable';
-
-const GET_PEAK_LIST = gql`
-  query getPeakList($id: ID!, $userId: ID!) {
-    peakList(id: $id) {
-      id
-      name
-      shortName
-      type
-      mountains {
-        id
-        name
-        latitude
-        longitude
-        elevation
-        prominence
-        state {
-          id
-          name
-          regions {
-            id
-            name
-            states {
-              id
-            }
-          }
-        }
-      }
-      parent {
-        id
-        mountains {
-          id
-          name
-          latitude
-          longitude
-          elevation
-          prominence
-          state {
-            id
-            name
-            regions {
-              id
-              name
-              states {
-                id
-              }
-            }
-          }
-        }
-      }
-    }
-    user(id: $userId) {
-      id
-      name
-      peakLists {
-        id
-      }
-      mountains {
-        mountain {
-          id
-        }
-        dates
-      }
-    }
-  }
-`;
-
-export interface MountainDatum {
-  id: Mountain['id'];
-  name: Mountain['name'];
-  latitude: Mountain['latitude'];
-  longitude: Mountain['longitude'];
-  elevation: Mountain['elevation'];
-  prominence: Mountain['prominence'];
-  state: {
-    id: State['id'];
-    name: State['name'];
-    regions: Array<{
-      id: Region['id'];
-      name: Region['name'];
-      states: Array<{
-        id: State['id'],
-      }>
-    }>
-  };
-}
-
-export interface PeakListDatum {
-  id: PeakList['id'];
-  name: PeakList['name'];
-  shortName: PeakList['shortName'];
-  type: PeakList['type'];
-  mountains: MountainDatum[] | null;
-  parent: {
-    id: PeakList['id'];
-    mountains: MountainDatum[] | null;
-  } | null;
-}
-
-export interface UserDatum {
-  id: User['id'];
-  name: User['name'];
-  peakLists: Array<{
-    id: PeakList['id'];
-  }>;
-  mountains: User['mountains'];
-}
-
-interface SuccessResponse {
-  peakList: PeakListDatum;
-  user: UserDatum;
-}
-
-interface Variables {
-  id: string;
-  userId: string;
-}
+import MountainDetail from '../../mountains/detail/MountainDetail';
+import PeakListDetail from './PeakListDetail';
 
 interface Props extends RouteComponentProps {
   userId: string;
@@ -133,57 +14,30 @@ interface Props extends RouteComponentProps {
 
 const PeakListDetailPage = (props: Props) => {
   const { userId, match } = props;
-  const { id }: any = match.params;
+  const { id, mountainId }: any = match.params;
 
-  const {loading, error, data} = useQuery<SuccessResponse, Variables>(GET_PEAK_LIST, {
-    variables: { id, userId },
-  });
-  if (loading === true) {
-    return null;
-  } else if (error !== undefined) {
-    console.error(error);
-    return null;
-  } else if (data !== undefined) {
-    const {
-      peakList: {parent, type},
-      peakList, user,
-    } = data;
-    let mountains: MountainDatum[];
-    if (parent !== null && parent.mountains !== null) {
-      mountains = parent.mountains;
-    } else if (peakList.mountains !== null) {
-      mountains = peakList.mountains;
-    } else {
-      mountains = [];
-    }
-    const completedAscents = user.mountains !== null ? user.mountains : [];
-    return (
-      <>
-        <ContentLeftLarge>
-          <ContentBody>
-            <Header
-              user={user}
-              mountains={mountains}
-              peakList={peakList}
-              completedAscents={completedAscents}
-            />
-            <MountainTable
-              user={user}
-              mountains={mountains}
-              type={type}
-            />
-          </ContentBody>
-        </ContentLeftLarge>
-        <ContentRightSmall>
-          <ContentBody>
-            selected mountain content
-          </ContentBody>
-        </ContentRightSmall>
-      </>
-    );
-  } else {
-    return null;
-  }
+  const mountainDetail = mountainId === undefined
+    ? (
+      <h2>Select a mountain to see more details</h2>
+      )
+    : (
+      <MountainDetail userId={userId} id={mountainId} />
+      );
+
+  return (
+    <>
+      <ContentLeftLarge>
+        <ContentBody>
+          <PeakListDetail userId={userId} id={id} />
+        </ContentBody>
+      </ContentLeftLarge>
+      <ContentRightSmall>
+        <ContentBody>
+          {mountainDetail}
+        </ContentBody>
+      </ContentRightSmall>
+    </>
+  );
 };
 
 export default withRouter(PeakListDetailPage);
