+++ client/src/routing/routes.ts
@@ -2,12 +2,17 @@ export enum Routes {
   Login = '/',
   Dashboard = '/',
   Lists = '/lists',
+  ListsWithDetail = '/lists/:id',
   ListDetail = '/list/:id',
+  ListDetailWithMountainDetail = '/list/:id/mountain/:mountainId',
   MountainDetail = '/mountain/:id',
   Friends = '/users',
+  FriendsWithProfile = '/users/:id',
   UserProfile = '/user/:id',
-  ComparePeakList = '/user/:friendId/list/:peakListId',
+  ComparePeakList = '/user/:id/list/:peakListId',
   CompareAllPeaks = '/user/:id/compare-all-peaks',
+  ComparePeakListIsolated = '/compare/user/:id/list/:peakListId',
+  CompareAllPeaksIsolated = '/compare/user/:id/compare-all-peaks',
   Admin = '/admin',
   AdminRegions = '/admin/regions',
   AdminStates = '/admin/states',
